### 2019-06-05 #311 @Robinlovelace
Presubmission enquiry: gitgantt
Submitting Author: Name (@robinlovelace)
Repository:   https://github.com/Robinlovelace/gitgantt

---

```
Package: gitgantt
Title: Create Gantt Charts with GitHub Issues and R
Version: 0.0.0.9000
   person("Robin", "Lovelace", email = "rob00x@gmail.com", role = c("aut", "cre"),
Description: Downloads GitHub issues that have GanttStart and GanttDue tags and converts them into Gantt charts.
    gh,
    timevis
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

Creating project plans is often a frustrating and non-reproducible process. This package can help ensure reproducibility, transparency and integration with community/team by linking a key part of a project plan with GitHub issues.

Anyone needing project time-lines, e.g. for funding applications, communicating planned work, communicating the status of scientific projects and packages.

See https://github.com/Robinlovelace/gitgantt#thanks

I have needed this package many times in the past an wasted hours creating gantt charts using many adhoc workflows, sometimes involving Microsoft Excel. I think it's an issue that really slows many projects down and leads to duplication of effort between GitHub tracking and 'client facing' plans. Finally I've developed something that works! I don't have a huge amount of time for maintenance, but suspect it could be very useful to many people, so thinking that if it's in a community-supported place like rOpenSci that could help me, and potential users, ensure it's robust and well maintained. Have already asked @mpadge and @layik to take a look, but think having eyes of a much larger community could greatly benefit this nascent project. Any thoughts on best way forward very welcome.
