### 2019-06-05 #311 @noamross
Thanks for the inquiry, @Robinlovelace! This looks useful (I may use it myself!), but it is out-of-scope for us.

You may want to comment on this proposal we have to clarify Aims and Scope by breaking up the "reproducibility" category into its specific parts: https://github.com/ropensci/software-review-meta/issues/81 .
