### 2019-06-05 #311 @Robinlovelace
Thanks @noamross, closing for now in that case. The idea is to get the word out there and invite feedback. Do you have any further advice on how to go about that? ATFutures, which also hosts the [`calendar`](https://github.com/ATFutures/calendar) package, is my default plan for a home for this. To anyone else who reads this: new issues/PRs are very welcome https://github.com/Robinlovelace/gitgantt/issues
