### 2019-06-05 #311 @noamross
I've long suggested that people use our [forum](https://discuss.ropensci.org/) to solicit feedback on packages/projects that weren't explicitly in our peer-review process.  In the past this has limited success, but I think that we finally have gotten enough people on the forum through other projects that there's a critical mass.  And maybe I can get @stefaniebutland and @sckott to add a "This week on the forum" to our newsletter and do something similar on Twitter/Slack 😉
