### 2019-06-05 #311 @maelle
You can also post the link to the package to R Weekly, and if you write a blog post somewhere explaining what the package does and where help is needed, that link too can go to R Weekly.
