### 2019-06-06 #311 @Robinlovelace
One follow-up on this: I posted this after seeing gitignore accepted as it aids reproducibility: https://github.com/ropensci/software-review/issues/303

The statements as to why it fits that category were "Because it helps facilitate use of version control." Thinking about it, I would say the aim of this package is similar: it aids the use of a version control and collaboration platform in a way that should encourage transparent (in that the plans will be visible publicly rather than just in private reports or funding applications) and reproducible (because the visuals can be reproduced, unlike most time planning software) work.

Clarification welcome, hope this helps rOpenSci process if nothing else.
