### 2019-06-06 #311 @Robinlovelace
Looking at https://github.com/ropensci/software-review-meta/issues/81  I just spotted:

> Scientific workflow automation and testing

Could it fit that category? I guess it depends whether project management and timeline generation is defined as part of a scientific workflow. `gitgantt` certainly automates it.
