### 2019-06-09 #311 @noamross
Thanks for your follow-up @Robinlovelace. Certainly there are grey areas in this and we're always adjusting and refining as we find new edge cases or priorities.  You didn't leave anything out of your justification - we consider qualification against all categories, using authors' justification as a place to start.  It's my opinon that **gitgantt** still falls out of scope, in that it primarily uses other features of GitHub than version control itself, and falls into scientific workflow only if we would include any report-generating package in our scope.

That said, I am hoping to take this for a spin soon - I think it would be neat to incorporate a project gantt chart right into it's website.
