### 2019-06-10 #311 @Robinlovelace
Hi @noamross thank you for the follow-up too. Makes sense. Gantt chart / timeline report generation is not something that is specific to *scientific* work. I put it together for one specific project. Will be stoked if it's of use for others. Any feedback very welcome. @mpadge has raised the possibility of going further, creating a project management tool that can read and write to GH issues but for now I think the priority is polishing what is already there, perhaps with a focus on nice default visuals.
