### 2019-06-10 #311 @maelle
If you two take it further have a look at https://github.com/emilyriederer/projmgr that has similar aims I think.
